move smartCompletionPrivateRegistryImage to Utils class
