--- azure-mgmt-appservice/src/test/resources/session-records/testWebAppPrivateRegistryImage.json
@@ -1,4 +1,0 @@
-{
-  "networkCallRecords" : [ ],
-  "variables" : [ "java-func-26e835488", "java-func-12e81778e", "java-func-bbb58603e", "java-asp-52a717702", "java-asp-39c792780", "javastorea0880619d", "javacsmrgae1227109", "javacsmrg3d442532e", "javacsmrg95828734b", "84161bf8-adb6-46d6-8fb0-b7537cdf1794", "444f5d74-970f-49e5-9462-2299718360bf", "java-func-26e835488plana3526302", "ba4f8aad-43c6-42e5-b694-0921be27fc04", "b797f4a0-5338-49b4-8fe7-5eb8120bc588", "aeaf2a5e-8ca5-472a-afb5-fc4bd6aa0217", "java-func-26e835488pland6f41256", "eeea1446-10a7-40a3-84da-db137d868cbf", "91b21307-2f26-4206-b585-51d874d237af", "53c8ada0-340b-44ae-9a02-de91b8794c6c", "java-func-26e835488plan65d04937", "fe08c0fe-fdd5-4c5f-8e1f-f48e82448750", "5263ba6e-6a4b-48af-ac32-df4efd9d8b66", "e9cc95b6-c1c9-4309-b7be-5f720651cab9", "java-func-26e835488pland4c19834", "41f82965-31a0-4b8f-9f92-628342898b6a", "36db7148-294a-448e-a85d-ea1139e4ebc1", "fa7501eb-9b8c-4bb8-b44e-2ec220fe491c", "java-func-26e835488plan4f299721", "78ebbeca-f65d-4cc8-a793-b955c46f2d80", "b158e4a2-6c7c-41e9-94ea-91f08cf3e249", "6af7845c-0929-47ad-9b39-a90d23475988", "java-func-26e835488plan7fb95637", "7c94c01a-b9da-48fb-8693-df27d6c959e7", "a37ff8fa-32a5-46b9-ad73-b0fe5f1aa970", "c060dc05-2ce2-4b26-a753-10f80e449ad7", "java-func-26e835488plan70288568", "7af1f0e1-3c76-4093-b8a1-4c4b617b54ea", "0c104e43-5e75-4afc-871c-958166bb9823", "46356758-6bca-470b-8526-4c15c8f8919e", "java-func-26e835488plan40d50462", "9431f648-9187-4d15-9967-483c59ab5750" ]
-}
